ProductSentimentAnalysis/airtag/pyproject.toml
[tool.poetry]
name = "airtag"
version = "0.1.0"
description = "Python Application Template Getting Started"
authors = ["PBG"]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
numpy = "^1.21.2"
pytest = "^6.2.5"
SQLAlchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
sqlalchemy-json = "^0.4.0"
requests = "^2.26.0"
pyaml = "^21.8.3"
pandas = "^1.3.3"
Unidecode = "^1.3.2"
tweepy = "^4.4.0"
matplotlib = "^3.5.1"
pytz = "^2021.3"
nltk = "^3.6.5"
tweet-preprocessor = "^0.6.0"
plotly = "^5.4.0"
wordcloud = "^1.8.1"
azure-ai-textanalytics = "^5.1.0"
textblob = "^0.17.1"
streamlit = "^1.3.0"
watchdog = "^2.1.6"
vaderSentiment = "^3.3.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
flake8-isort = "ˆ4.0.0"
flake8-builtins = "ˆ1.5.3"
autopep8 = "ˆ1.5.7"
pylint = "ˆ2.8.2"
bandit = "ˆ1.7.0"
black = "ˆ21.5b0"
mypy = "ˆ0.812"
flake8-bandit = "ˆ2.1.2"
flake8-mypy = "ˆ17.8.0"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
autoflake = "^1.3.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core=1.1.8"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "functions: mark test for the base functionalities",
    "core: all about the core",
    "db: all about the database",
    "api: all about the api of the system",
]